SPPM: let the compiler do its job by removing volatile
volatile is in general a bad idea and should only be used for pointer to device
or stuff like that.

In the case of the grid and jump_list data, adding a volatile attribute while
avoid possibles optimisations during the photonpass.

During the eyepass, the data is instead shared, but the memory shynchronisation
is handled by the atomic_cas32 function (that is their purpose).

See the discussion on http://stackoverflow.com/questions/4557979/when-to-use-volatile-with-multi-threading