SPPM: use the LuxRender api for halt SPP and halt time
This commit is a bit crappy. Mainly I'm waiting for a change in the film API
which will allows to have buffer which different sample count.

So now: Eye sample does not increase the photon count and Photon sampler
increase the photon count in a way that at the end of a pass, the photon count
equal the number of pixel of the screen (hence haltspp will work). The photon
sampler must take this change into account and normalize alpha differently.

--HG--
extra : rebase_source : eacdd6292c58f3d6f184d4b9b300e4c4c2853585