Updated luxcomp with RMS metric, full-film comparison and various improvements. Thanks to Zsolnai for patch.
To compliment the Mean Square Error metric, CompareFilmWith/4 can now be asked to compute the the Root Mean Square (RMS) metric.

We compute the difference between the reference and test film for comparison. Small-medium-big diff computations instead of small-big. Small diff now means within 10% of error, medium diff is between 10% and 50%, big diff is over 50%.

The magnitude of the minimum and maximum difference is now computed and displayed.

The average difference between the reference and test film is also computed and displayed.

The -t flag now will return the desired error metric (0 - MSE (default), 1 - RMS error). As MSE is given for the default metric, it won't break anything.

Beside per-buffer comparison, it is now possible to compare the error for rendered images.

Some cleanups here and there.