Fixed issue with slow network file transfer.
Disabled the naggle algorithm. Apparently the explicit flushes were not sufficient. This will lead to some increased packet traffic, but I think that on most modern networks that will not be an issue.

Also replaced some endl with \n, as endl calls flush and we don't need that everywhere.

Added exceptions instead of silently failing in case of protocol error to prevent corrupted rendering results.