Fixed issue where hybrid renderer would enter an invalid state when loading scenes with incompatible light strategy.
If the check for compatible light strategy fails, the hybrid renderer is still in the RUN state, but is not actually rendering.

Due to being in the RUN state, additional rendering threads may be added by the user (luxrender/luxconsole). Since the renderer isn't really in the running state, this causes issues.